minor refactoring - moved validation classes to different package - updated docs
